Move cast to u8 to a better spot

+++ src/board/mod.rs
@@ -182,12 +182,12 @@ impl Board {
                 rank -= 1;
                 file = 0;
             } else if c.is_digit(10) {
-                let offset = c.to_digit(10).unwrap() as u8;
+                let offset = c.to_digit(10).unwrap();
                 if file + offset > 8 {
                     return parse_fen_error!("Invalid fen string '{}': invalid offset {} at pos {}", fen_pos, offset, idx);
                 }
 
-                file += offset;
+                file += offset as u8;
             } else {
                 if file >= 8 {
                     return parse_fen_error!("Invalid fen string '{}': position goes out of bounds at pos {}", fen_pos, idx)
